+++ crates/ruma-events/src/enums.rs
@@ -8,6 +8,20 @@ use serde_json::value::RawValue as RawJsonValue;
 
 use super::room::encrypted;
 
+/// Event types that servers should send as [stripped state] to help clients identify a room when
+/// they can't access the full room state.
+///
+/// [stripped state]: https://spec.matrix.org/latest/client-server-api/#stripped-state
+pub const RECOMMENDED_STRIPPED_STATE_EVENT_TYPES: &[StateEventType] = &[
+    StateEventType::RoomCreate,
+    StateEventType::RoomName,
+    StateEventType::RoomAvatar,
+    StateEventType::RoomTopic,
+    StateEventType::RoomJoinRules,
+    StateEventType::RoomCanonicalAlias,
+    StateEventType::RoomEncryption,
+];
+
 event_enum! {
     /// Any global account data event.
     enum GlobalAccountData {